+++ flow-typed/File.js
@@ -76,12 +76,3 @@ declare type DeletePurchasedUri = {
     uri: string,
   },
 };
-
-declare type FileUploadItem = {
-  params: UpdatePublishFormData,
-  file: File,
-  fileFingerprint: string,
-  progress: string,
-  status?: string,
-  tusUploader?: any,
-};
