+++ ui/constants/action_types.js
@@ -331,6 +331,9 @@ export const PUBLISH_FAIL = 'PUBLISH_FAIL';
 export const CLEAR_PUBLISH_ERROR = 'CLEAR_PUBLISH_ERROR';
 export const REMOVE_PENDING_PUBLISH = 'REMOVE_PENDING_PUBLISH';
 export const DO_PREPARE_EDIT = 'DO_PREPARE_EDIT';
+export const UPDATE_UPLOAD_ADD = 'UPDATE_UPLOAD_ADD';
+export const UPDATE_UPLOAD_PROGRESS = 'UPDATE_UPLOAD_PROGRESS';
+export const UPDATE_UPLOAD_REMOVE = 'UPDATE_UPLOAD_REMOVE';
 
 // Media
 export const MEDIA_PLAY = 'MEDIA_PLAY';
@@ -490,11 +493,6 @@ export const FETCH_VIEW_COUNT_COMPLETED = 'FETCH_VIEW_COUNT_COMPLETED';
 export const FETCH_SUB_COUNT_STARTED = 'FETCH_SUB_COUNT_STARTED';
 export const FETCH_SUB_COUNT_FAILED = 'FETCH_SUB_COUNT_FAILED';
 export const FETCH_SUB_COUNT_COMPLETED = 'FETCH_SUB_COUNT_COMPLETED';
-
-// Lbry.tv
-export const UPDATE_UPLOAD_ADD = 'UPDATE_UPLOAD_ADD';
-export const UPDATE_UPLOAD_PROGRESS = 'UPDATE_UPLOAD_PROGRESS';
-export const UPDATE_UPLOAD_REMOVE = 'UPDATE_UPLOAD_REMOVE';
 
 // User
 export const GENERATE_AUTH_TOKEN_FAILURE = 'GENERATE_AUTH_TOKEN_FAILURE';
