+++ ui/page/fileListPublished/index.js
@@ -5,9 +5,9 @@ import {
   selectMyClaimsPageItemCount,
   selectFetchingMyClaimsPageError,
 } from 'redux/selectors/claims';
+import { selectUploadCount } from 'redux/selectors/publish';
 import { doFetchClaimListMine, doCheckPendingClaims } from 'redux/actions/claims';
 import { doClearPublish } from 'redux/actions/publish';
-import { selectUploadCount } from 'lbryinc';
 import FileListPublished from './view';
 import { withRouter } from 'react-router';
 import { MY_CLAIMS_PAGE_SIZE, PAGE_PARAM, PAGE_SIZE_PARAM } from 'constants/claim';
