+++ ui/redux/actions/publish.js
@@ -688,3 +688,31 @@ export const doCheckReflectingFiles = () => (dispatch: Dispatch, getState: GetSt
     }, 5000);
   }
 };
+
+export function doUpdateUploadAdd(file: File | string, params: { [key: string]: any }, tusUploader: any) {
+  return (dispatch: Dispatch, getState: GetState) => {
+    dispatch({
+      type: ACTIONS.UPDATE_UPLOAD_ADD,
+      data: { file, params, tusUploader },
+    });
+  };
+}
+
+export const doUpdateUploadProgress = (props: {
+  params: { [key: string]: any },
+  progress?: string,
+  status?: string,
+}) => (dispatch: Dispatch) =>
+  dispatch({
+    type: ACTIONS.UPDATE_UPLOAD_PROGRESS,
+    data: props,
+  });
+
+export function doUpdateUploadRemove(params: { [key: string]: any }) {
+  return (dispatch: Dispatch, getState: GetState) => {
+    dispatch({
+      type: ACTIONS.UPDATE_UPLOAD_REMOVE,
+      data: { params },
+    });
+  };
+}
