+++ ui/redux/reducers/publish.js
@@ -1,9 +1,12 @@
 // @flow
 import { handleActions } from 'util/redux-utils';
 import { buildURI } from 'util/lbryURI';
+import { serializeFileObj } from 'util/file';
 import * as ACTIONS from 'constants/action_types';
 import * as THUMBNAIL_STATUSES from 'constants/thumbnail_upload_statuses';
 import { CHANNEL_ANONYMOUS } from 'constants/claim';
+
+const getKeyFromParam = (params) => `${params.name}#${params.channel || 'anonymous'}`;
 
 type PublishState = {
   editingURI: ?string,
@@ -38,6 +41,7 @@ type PublishState = {
   tags: Array<string>,
   optimize: boolean,
   useLBRYUploader: boolean,
+  currentUploads: { [key: string]: FileUploadItem },
 };
 
 const defaultState: PublishState = {
@@ -78,6 +82,7 @@ const defaultState: PublishState = {
   publishError: undefined,
   optimize: false,
   useLBRYUploader: false,
+  currentUploads: {},
 };
 
 export const publishReducer = handleActions(
@@ -96,6 +101,7 @@ export const publishReducer = handleActions(
       bid: state.bid,
       optimize: state.optimize,
       language: state.language,
+      currentUploads: state.currentUploads,
     }),
     [ACTIONS.PUBLISH_START]: (state: PublishState): PublishState => ({
       ...state,
@@ -127,7 +133,65 @@ export const publishReducer = handleActions(
         ...publishData,
         editingURI: uri,
         uri: shortUri,
+        currentUploads: state.currentUploads,
       };
+    },
+    [ACTIONS.UPDATE_UPLOAD_ADD]: (state: PublishState, action) => {
+      const { file, params, tusUploader } = action.data;
+      const key = getKeyFromParam(params);
+      const currentUploads = Object.assign({}, state.currentUploads);
+
+      currentUploads[key] = {
+        file,
+        fileFingerprint: serializeFileObj(file), // TODO: get hash instead?
+        progress: '0',
+        params,
+        tusUploader,
+      };
+
+      return { ...state, currentUploads };
+    },
+    [ACTIONS.UPDATE_UPLOAD_PROGRESS]: (state: PublishState, action) => {
+      const { params, progress, status } = action.data;
+      const key = getKeyFromParam(params);
+      const currentUploads = Object.assign({}, state.currentUploads);
+
+      if (progress) {
+        currentUploads[key].progress = progress;
+        delete currentUploads[key].status;
+      } else if (status) {
+        currentUploads[key].status = status;
+        if (status === 'error') {
+          delete currentUploads[key].tusUploader;
+        }
+      }
+
+      return { ...state, currentUploads };
+    },
+    [ACTIONS.UPDATE_UPLOAD_REMOVE]: (state: PublishState, action) => {
+      const { params } = action.data;
+      const key = getKeyFromParam(params);
+      const currentUploads = Object.assign({}, state.currentUploads);
+
+      delete currentUploads[key];
+
+      return { ...state, currentUploads };
+    },
+    [ACTIONS.REHYDRATE]: (state: PublishState, action) => {
+      if (action && action.payload && action.payload.publish) {
+        const newPublish = { ...action.payload.publish };
+
+        // Cleanup for 'publish::currentUploads'
+        if (newPublish.currentUploads) {
+          Object.keys(newPublish.currentUploads).forEach((key) => {
+            delete newPublish.currentUploads[key].tusUploader;
+          });
+        }
+
+        return newPublish;
+      }
+
+      return state;
     },
   },
   defaultState
