+++ ui/redux/selectors/publish.js
@@ -119,3 +119,10 @@ export const selectTakeOverAmount = createSelector(
     return null;
   }
 );
+
+export const selectCurrentUploads = (state) => selectState(state).currentUploads;
+
+export const selectUploadCount = createSelector(
+  selectCurrentUploads,
+  (currentUploads) => currentUploads && Object.keys(currentUploads).length
+);
